Add tests and incorporate code review comments
